pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anvil-tui"
version = "0.1.1"
description = "Terminal User Interface for managing Microsoft Foundry projects and resources"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Kristofer", email = "kristofer@example.com" }
]
keywords = ["tui", "terminal", "microsoft", "foundry", "azure", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: User Interfaces",
    "Typing :: Typed",
]
dependencies = [
    "textual>=0.89.0",
    "httpx>=0.28.0",
    "azure-identity>=1.19.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-cognitiveservices>=13.7.0",
    "azure-ai-projects>=2.0.0b2",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "textual-dev>=1.7.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
anvil = "anvil.app:main"
anvil-tui = "anvil.app:main"

[project.urls]
Homepage = "https://github.com/mklab-se/anvil"
Repository = "https://github.com/mklab-se/anvil"
Issues = "https://github.com/mklab-se/anvil/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/anvil"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["anvil"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --cov=anvil --cov-report=term-missing"

[tool.coverage.run]
source = ["src/anvil"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
